proxy/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websocket-client==1.6.4
pydantic==2.5.0
python-multipart==0.0.6
httpx==0.25.2
